requirements.txt
anthropic>=0.40.0
flask>=3.0.0
gunicorn>=21.0.0
streamlit>=1.40.0
pydantic>=2.0.0
pytrends>=4.9.2
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
